x-ops/requirements.txt
PyYAML==6.0.1
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
playwright==1.50.0
